+++ modules/Sources/Features/CoordFlows/SignWithKeystoneCoordFlowView.swift
@@ -41,6 +41,8 @@ public struct SignWithKeystoneCoordFlowView: View {
                 .navigationBarHidden(true)
             } destination: { store in
                 switch store.case {
+                case let .keystoneFirmwareUpdate(store):
+                    KeystoneFirmwareUpdateView(store: store)
                 case let .preSendingFailure(store):
                     PreSendingFailureView(store: store, tokenName: tokenName)
                 case let .scan(store):