+++ modules/Sources/Features/CoordFlows/SwapAndPayCoordFlowStore.swift
@@ -34,6 +34,7 @@ public struct SwapAndPayCoordFlow {
         case addressBookContact(AddressBook)
         case confirmWithKeystone(SendConfirmation)
         case crossPayConfirmation(SwapAndPay)
+        case keystoneFirmwareUpdate(SendConfirmation)
         case preSendingFailure(SendConfirmation)
         case scan(Scan)
         case sending(SendConfirmation)